feat: use profileCard for displaying users' profiles
